+++ src/main/java/dev/enjarai/trickster/item/ModItems.java
@@ -6,14 +6,18 @@ import dev.enjarai.trickster.item.component.SelectedSlotComponent;
 import dev.enjarai.trickster.item.component.SpellComponent;
 import dev.enjarai.trickster.spell.SpellPart;
 import io.wispforest.lavender.book.LavenderBookItem;
+import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroup;
+import net.minecraft.block.Block;
 import net.minecraft.component.DataComponentTypes;
 import net.minecraft.component.type.ContainerComponent;
 import net.minecraft.item.Item;
+import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.registry.RegistryKeys;
 import net.minecraft.registry.tag.TagKey;
+import net.minecraft.text.Text;
 import net.minecraft.util.collection.DefaultedList;
 
 public class ModItems {
@@ -37,12 +41,25 @@ public class ModItems {
     public static final TagKey<Item> CAN_EVALUATE_DYNAMICALLY = TagKey.of(RegistryKeys.ITEM, Trickster.id("can_evaluate_dynamically"));
     public static final TagKey<Item> HOLDABLE_HAT = TagKey.of(RegistryKeys.ITEM, Trickster.id("holdable_hat"));
     public static final TagKey<Item> SCROLLS = TagKey.of(RegistryKeys.ITEM, Trickster.id("scrolls"));
+    public static final TagKey<Block> CONJURABLE_FLOWERS = TagKey.of(RegistryKeys.BLOCK, Trickster.id("conjurable_flowers"));
+
+    public static final ItemGroup ITEM_GROUP = FabricItemGroup.builder()
+            .icon(WAND::getDefaultStack)
+            .displayName(Text.translatable("trickster.item_group"))
+            .entries((context, entries) -> {
+                entries.add(TOME_OF_TOMFOOLERY);
+                entries.add(SCROLL_AND_QUILL);
+                entries.add(MIRROR_OF_EVALUATION);
+                entries.add(TOP_HAT);
+                entries.add(WAND);
+            })
+            .build();
 
     private static <T extends Item> T register(String name, T item) {
         return Registry.register(Registries.ITEM, Trickster.id(name), item);
     }
 
     public static void register() {
-
+        Registry.register(Registries.ITEM_GROUP, Trickster.id("trickster"), ITEM_GROUP);
     }
 }